% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_table.r
\name{extract_tables}
\alias{extract_tables}
\title{Function to extract tables}
\usage{
extract_tables(text_lines, delimiter = "\\\\s{2,}",
  delimiter_table = "\\\\s{2,}", replacement = "\\\\/",
  col_names = FALSE)
}
\arguments{
\item{text_lines}{The text of a PDF document read into R and split into 
lines.}

\item{delimiter}{A delimiter used to detect tables. The default is two 
consecutive blank white spaces.}

\item{delimiter_table}{A delimiter used to separate table cells. The default
value is two consecutive blank white spaces.}

\item{replacement}{A delimiter used to separate table cells after the 
replacement of white space is done.}

\item{col_names}{TRUE/FALSE value passed to `readr::read_delim` 
to indicate if column names should be used. Default value is FALSE which 
means column names will be generic (i.e. X1, X2, etc). A value of TRUE
would take the values from the first row of data extracted.}
}
\description{
Function to extract tables
}
