% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_pdf.r
\name{split_pdf}
\alias{split_pdf}
\title{Split multicolumn pdfs}
\usage{
split_pdf(x, numcols = 2, delim = " {2,}")
}
\arguments{
\item{x}{A list object already processed via pdf_text.}

\item{numcols}{Number of columns in the pdf, default is 2.}

\item{delim}{The delimiter to split multicolumns. The default is three spaces, 
changing this delimiter can be useful if pdf has not been split correctly.}
}
\description{
The \code{\link{pdf_text}} function does not handle multicolumn pdfs well,
it simply returns the entire line spanning across multiple columns. This function
will split multiple columns and retain their place in the pdf.
}

