% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_search.r
\name{keyword_directory}
\alias{keyword_directory}
\title{Wrapper for keyword search function}
\usage{
keyword_directory(directory, keyword, surround_lines = FALSE,
  ignore.case = FALSE, full_names = FALSE, recursive = FALSE)
}
\arguments{
\item{directory}{The directory to perform the search for pdf files to search.}

\item{keyword}{The keyword(s) to be used to search in the text. Multiple 
keywords can be specified with a character vector.}

\item{surround_lines}{numeric/FALSE indicating whether the output should extract the 
surrouding lines of text in addition to the matching line. Default is FALSE, if 
not false, include a numeric number that indicates the additional number of 
surrounding lines that will be extracted.}

\item{ignore.case}{TRUE/FALSE indicating whether the case of the keyword matters. 
Default is FALSE meaning that case of the keyword is literal.}

\item{full_names}{TRUE/FALSE indicating if the full file path should be used.
Default is FALSE, see \code{\link{list.files}} for more details.}

\item{recursive}{TRUE/FALSE indicating if subdirectories should be searched as well.
Default is FALSE, see \code{\link{list.files}} for more details.}
}
\description{
This function allows one to loop over pdf files within a single directory
}

